requirements.txt
fastapi>=0.115.0
uvicorn[standard]
sqlalchemy[asyncio]>=2.0.0
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
python-dotenv
httpx
pytest
pytest-asyncio
pytest-cov
rich
